fix_api(be): Fixed all use cases related to cost and consumption and added migration to change that fields from decimal to double

+++ backend/ElectroManage.Application/DTO's/ConsumptionDTO.cs
@@ -5,5 +5,5 @@ public record ConsumptionDTO
     public long CompanyId { get; set; }
     public string CompanyName { get; set; } = null!;
     public decimal ConsumptionLimit { get; set; }
-    public decimal TotalConsumption { get; set; }
+    public double TotalConsumption { get; set; }
 }
